Drop alias_dist module in favor of rand crate
The `rand` crate now has an alias_method::WeightedIndex distribution
that does exactly what my custom alias_dist module was doing.  Since I
already depend on the crate, and it's very well maintained, it seems
silly to continue to maintain my own version.

Performance seems to be unchanged for practical purposes. I also
tried the non-alias-method WeightedIndex. The benchmark was a couple
percent slower in that case, so it's nice to see that using the alias
method wasn't just a waste of time.